% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_importance.R
\name{get_importance}
\alias{get_importance}
\title{Collect Variable Importance Metrics for Covariates Used in Random Forest Adjustment}
\usage{
get_importance(rfa)
}
\arguments{
\item{rfa}{an `rfa()` fitted object}
}
\description{
This function returns variable importance metrics (either "impurity" or "permutation") for each of the covariates
included in random forest adjustment estimation with respect to the response
variable and the explanatory variable of interest.
}
\details{
To return importance metrics, in `rfa()` the argument `importance` must be set to either "impurity" or "permutation" (quoted).
}
