\name{northCascades}
\alias{northCascades}
\title{Data set to perform a Regional Frequency Analysis}
\usage{data(northCascades)}
\description{
 The data set of Hosking LMOMENTS Package.
}
\format{A dataframe which include the names of sites, sample size, and
  samples L-CV, L-Skewness and L-Kurtosis for each site.}
\keyword{datasets}
