% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{plot.rfa}
\alias{plot.rfa}
\title{Wrapper for ggplot2 to Plot Histogram of Empirical ATE Replicates}
\usage{
\method{plot}{rfa}(model)
}
\description{
This function takes output from a `rfa` object and returns a
histogram of the bootstrapped ATE replicates.
}
