% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rfa.R
\name{plot_rfa}
\alias{plot_rfa}
\title{Wrapper for ggplot2 to Plot Histogram of Empirical ATE Replicates}
\usage{
plot_rfa(model, varname = "predictor")
}
\description{
This function takes output from a `rfa` object and returns a
histogram of the bootstrapped ATE replicates.
}
