\name{Data set}
\alias{region}
\title{A data set to perform a Regional Frequency Analysis}
\usage{data(region)}
\description{
  This is a list containing high flood events on 15 French gauging
  stations. The list also contains one more argument \code{'record'}
  which specifies the record length for each site.}
\format{A list with 16 arguments.}
\keyword{datasets}