% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa.R
\name{rfa}
\alias{rfa}
\title{Estimate Average Treatment Effect via Random Forest Adjustment}
\usage{
rfa(formula, data = NULL)
}
\arguments{
\item{formula}{a formula object where the first right-hand side variable
is the predictor variable of interest (or the treatment variable) and
the remaining right-hand side predictors are confounding variables.}

\item{data}{an optional data frame containing the variables used to
implement the RFA routine.}
}
\value{
`rfa` returns a list containing the estimated ATE
and a vector of bootstrapped ATE estimates (useful if the user wishes)
to visualize the empirical distribution of the ATE. It also returns
a data frame of the predictor variable and response variable, both
pre- and post- adjustment.
}
\description{
This function estimates the average treatment effect of an explanatory
variable on some response variable using a procedure called Random Forest
Adjustment (RFA). The function accepts a formula object and dataframe
as inputs. rfa() assumes that the first right-hand side variable in the
formula object is the explanatory variable of interest and that all other
variables on the right-hand side are confounding variables used to
residualize the predictor and response prior to estimating the ATE. NA
values are allowed.
}
\details{
The RFA routine provides an estimate of the ATE of some
predictor (either binary or continuous) on a response variable, adjusting
for the confounding influence of other variables. As its name implies,
RFA substracts away the variance in the treatment and response explained
by confounding variables via random forest regression, prior to
estimating the ATE of the treatment on the response.
}
