% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rfa.R
\name{summary.rfa}
\alias{summary.rfa}
\title{Return Summary Statistics for ATE Estimated with `rfa`}
\usage{
\method{summary}{rfa}(model)
}
\arguments{
\item{model}{a `rfa` model object}
}
\value{
Returns summary statistics for the ATE obtained via the
RFA routine.
}
\description{
The function returns the estimated ATE, standard error,
t-statistic, and p-value for a `rfa` model object.
}
\details{
The function takes the estimated ATE and bootstrapped estimates
returned by an `rfa` object to calculate the summary statistics for the
ATE calculated from the RFA routine.
}
\seealso{
The `rfa` function.
}
