% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_rfa.R
\name{summary_rfa}
\alias{summary_rfa}
\title{Return Summary Statistics for `rfa` object.}
\usage{
summary_rfa(model)
}
\arguments{
\item{model}{a `rfa` model object}
}
\value{
Returns summary statistics for the ATE obtained via the
RFA routine.
}
\description{
The function returns the estimated coefficient, standard error,
t-statistic, p-value, confidence intervals, and random forest fit
performance for a `rfa` model object.
}
\seealso{
The `rfa` function.
}
