% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfa.R
\name{rfa}
\alias{rfa}
\title{Estimate Average Treatment Effect via Random Forest Adjustment}
\usage{
rfa(formula, covariates, data = NULL, se_type = "stata", clusters = NULL, ...)
}
\arguments{
\item{formula}{a formula object where the left-hand variable is the outcome
and the right-hand variable is the explanatory variable of interest.}

\item{covariates}{a formula object only containing the right-hand side specifying
the covariates to be used in the random forest regressions.}

\item{data}{an optional data frame containing the variables used to
implement the RFA routine.}

\item{se_type}{specifies the standard errors to be returned. If `clusters` is not
specified, the user can specify "classical", "HC0", "stata" (equivalent to "HC1"),
"HC2", or "HC3". If `clusters` is specified, the options are "CR0", "stata" (CR1),
and "CR2". "stata" is the default.}

\item{clusters}{optional name (quoted) of variable that corresponds to clusters
in the data.}

\item{...}{additional commands to override the default settings for implementing
random forests via `ranger`. See the `ranger` package for more details.}
}
\value{
`rfa` returns a list containing the model object (`lm_robust` object from
the `estimatr`` package), the data used to estimate the the model, a covariates matrix, and the
random forest regressions for the response and explanatory variable
(`ranger` objects from the `ranger` package).
}
\description{
This function estimates the average treatment effect of an explanatory
variable on some response variable using a procedure called Random Forest
Adjustment (RFA). RFA partials out the variation in a response and explanatory
variable of interest as a function of a set of covariates using random forest
regression. The latest version relies on the ranger package, which
provides a fast implementation of the random forest aglorithm (Brieman 2001).
}
\details{
The function accepts a formula object and dataframe
as inputs. rfa() assumes that the first right-hand side variable in the
formula object is the explanatory variable of interest and that all other
variables on the right-hand side are confounding variables used to
residualize the predictor and response prior to estimating the ATE. NA
values are allowed.

The RFA routine provides an estimate of the marginal relationship
between some
predictor (either binary or continuous) and a response variable, adjusting
for the confounding influence of other variables. As its name implies,
RFA substracts away the variance in the predictor of interest and response explained
by confounding variables via random forest regression.
}
